+++ src/fastermsc.rs
@@ -470,7 +470,7 @@ pub(crate) fn do_swap_k2<M, N, L>(
 #[cfg(test)]
 mod tests {
 	// TODO: use a larger, much more interesting example.
-	use crate::{arrayadapter::LowerTriangle, fastermsc, silhouette, util::assert_array};
+	use crate::{arrayadapter::LowerTriangle, fastermsc, silhouette, medoid_silhouette, util::assert_array};
 
 	#[test]
 	fn testfastermsc_simple() {
@@ -481,8 +481,10 @@ mod tests {
 		let mut meds = vec![0, 1, 2];
 		let (loss, assi, n_iter, n_swap): (f64, _, _, _) = fastermsc(&data, &mut meds, 10);
 		let (sil, _): (f64, _) = silhouette(&data, &assi, false);
+		let (msil, _): (f64, _) = medoid_silhouette(&data, &meds, false);
 		print!("FasterMSC: {:?} {:?} {:?} {:?} {:?} {:?}", loss, n_iter, n_swap, sil, assi, meds);
 		assert_eq!(loss, 0.9047619047619048, "loss not as expected");
+		assert_eq!(msil, 0.9047619047619048, "Medoid Silhouette not as expected");
 		assert_eq!(n_swap, 1, "swaps not as expected");
 		assert_eq!(n_iter, 2, "iterations not as expected");
 		assert_array(assi, vec![0, 0, 2, 1, 1], "assignment not as expected");
@@ -499,8 +501,10 @@ mod tests {
 		let mut meds = vec![0, 1];
 		let (loss, assi, n_iter, n_swap): (f64, _, _, _) = fastermsc(&data, &mut meds, 10);
 		let (sil, _): (f64, _) = silhouette(&data, &assi, false);
+		let (msil, _): (f64, _) = medoid_silhouette(&data, &meds, false);
 		print!("FasterMSC: {:?} {:?} {:?} {:?} {:?} {:?}", loss, n_iter, n_swap, sil, assi, meds);
 		assert_eq!(loss, 0.8805555555555555, "loss not as expected");
+		assert_eq!(msil, 0.8805555555555555, "Medoid Silhouette not as expected");
 		assert_eq!(n_swap, 3, "swaps not as expected");
 		assert_eq!(n_iter, 2, "iterations not as expected");
 		assert_array(assi, vec![0, 0, 0, 1, 1], "assignment not as expected");